requirement.txt
pygetwindow
pywin32
Pillow
numpy
pyautogui
Levenshtein
keyboard
opencv-python
paddlepaddle
pytesseract
PaddleOCR
